# packages/api-spec/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/play-invites": {
      "post": {
        "operationId": "playInvitesSend",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendInviteBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayInviteDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "\"Play together\" with a connection. 403 not_connected, 409 invite_pending,\n429 daily_limit (20 a day).",
        "tags": [
          "play"
        ]
      },
      "get": {
        "operationId": "playInvitesList",
        "parameters": [
          {
            "name": "direction",
            "required": false,
            "in": "query",
            "schema": {
              "default": "incoming",
              "type": "string",
              "enum": [
                "incoming",
                "outgoing"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayInviteListDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Pending invites to you (the banner) or from you.",
        "tags": [
          "play"
        ]
      }
    },
    "/v1/play-invites/{id}/accept": {
      "post": {
        "operationId": "playInvitesAccept",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayInviteDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Schedules the check-ins. 409 invite_expired after startsAt + 15 min.",
        "tags": [
          "play"
        ]
      }
    },
    "/v1/play-invites/{id}/decline": {
      "post": {
        "operationId": "playInvitesDecline",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "play"
        ]
      }
    },
    "/v1/me/check-ins": {
      "get": {
        "operationId": "checkInsDue",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckInListDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "\"Did you play with …?\" cards for Home.",
        "tags": [
          "play"
        ]
      }
    },
    "/v1/check-ins/{id}": {
      "post": {
        "operationId": "checkInsAnswer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerCheckInBody"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "One tap. The product's main number (docs/PRD.md).",
        "tags": [
          "play"
        ]
      }
    }
  },
  "info": {
    "title": "gamer-network API",
    "description": "Conventions: docs/API.md. Every non-2xx response is an ErrorResponse.",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PlayerStatus": {
        "type": "string",
        "enum": [
          "looking",
          "playing",
          "available_tonight",
          "not_available"
        ],
        "description": "Already reset to not_available once its time has passed."
      },
      "GameRefDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "shortCode": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "shortCode"
        ]
      },
      "RankRefDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "tier": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "name",
          "tier"
        ]
      },
      "RoleRefDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "UserGameDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "game": {
            "nullable": true,
            "description": "Null for a custom game.",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/GameRefDto"
              }
            ]
          },
          "customGameName": {
            "type": "string",
            "nullable": true,
            "description": "Set only for a custom game."
          },
          "rank": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/RankRefDto"
              }
            ]
          },
          "rankText": {
            "type": "string",
            "nullable": true,
            "description": "Free-text level, custom games only."
          },
          "role": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/RoleRefDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "game",
          "customGameName",
          "rank",
          "rankText",
          "role"
        ]
      },
      "UserCardDto": {
        "type": "object",
        "properties": {
          "status": {
            "description": "Already reset to not_available once its time has passed.",
            "allOf": [
              {
                "$ref": "#/components/schemas/PlayerStatus"
              }
            ]
          },
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string",
            "nullable": true
          },
          "city": {
            "type": "string"
          },
          "statusGame": {
            "nullable": true,
            "description": "The game behind \"Playing …\".",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/GameRefDto"
              }
            ]
          },
          "topGame": {
            "nullable": true,
            "description": "First game on the profile.",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserGameDto"
              }
            ]
          }
        },
        "required": [
          "status",
          "id",
          "username",
          "displayName",
          "avatarUrl",
          "city",
          "statusGame",
          "topGame"
        ]
      },
...
      "InviteWhen": {
        "type": "string",
        "enum": [
          "now",
          "in_30_min",
          "tonight"
        ]
      },
      "SendInviteBody": {
        "type": "object",
        "properties": {
          "playWhen": {
            "allOf": [
              {
                "$ref": "#/components/schemas/InviteWhen"
              }
            ]
          },
          "toUserId": {
            "type": "string",
            "description": "One of your connections.",
            "format": "uuid"
          },
          "gameId": {
            "type": "string",
            "description": "Any catalog game."
          }
        },
        "required": [
          "playWhen",
          "toUserId",
          "gameId"
        ]
      },
      "InviteStatus": {
        "type": "string",
        "enum": [
          "pending",
          "accepted",
          "declined",
          "expired"
        ],
        "description": "expired as soon as expiresAt passes while pending."
      },
      "PlayInviteDto": {
        "type": "object",
        "properties": {
          "playWhen": {
            "allOf": [
              {
                "$ref": "#/components/schemas/InviteWhen"
              }
            ]
          },
          "status": {
            "description": "expired as soon as expiresAt passes while pending.",
            "allOf": [
              {
                "$ref": "#/components/schemas/InviteStatus"
              }
            ]
          },
          "id": {
            "type": "string"
          },
          "from": {
            "$ref": "#/components/schemas/UserCardDto"
          },
          "to": {
            "$ref": "#/components/schemas/UserCardDto"
          },
          "game": {
            "$ref": "#/components/schemas/GameRefDto"
          },
          "startsAt": {
            "format": "date-time",
            "type": "string"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string",
            "description": "Count down to this; unanswered by then, it is expired."
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "playWhen",
          "status",
          "id",
          "from",
          "to",
          "game",
          "startsAt",
          "expiresAt",
          "createdAt"
        ]
      },
      "PlayInviteListDto": {
        "type": "object",
        "properties": {
          "items": {
            "description": "Pending and not expired, newest first.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlayInviteDto"
            }
          }
        },
        "required": [
          "items"
        ]
      },
      "CheckInDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "other": {
            "description": "Who you were matched with.",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserCardDto"
              }
            ]
          },
          "game": {
            "nullable": true,
            "description": "The game of the listing or invite that led here.",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/GameRefDto"
              }
            ]
          },
          "listingId": {
            "type": "string",
            "nullable": true
          },
          "playInviteId": {
            "type": "string",
            "nullable": true
          },
          "dueAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "other",
          "game",
          "listingId",
          "playInviteId",
          "dueAt"
        ]
      },
      "CheckInListDto": {
        "type": "object",
        "properties": {
          "items": {
            "description": "Due and unanswered, oldest first.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CheckInDto"
            }
          }
        },
        "required": [
          "items"
        ]
      },
      "CheckInAnswer": {
        "type": "string",
        "enum": [
          "played",
          "not_yet",
          "no"
        ]
      },
      "AnswerCheckInBody": {
        "type": "object",
        "properties": {
          "answer": {
            "allOf": [
              {
                "$ref": "#/components/schemas/CheckInAnswer"
              }
            ]
          }
        },
        "required": [
          "answer"
        ]
      },
...
    }
  }
}